modified files as per requests
