+++ tests/fixtures/reporters/watch_integration.py
@@ -1,0 +1,20 @@
+"""This script serves as the entry point for an integration test of the PyTA HTML report server.
+It is invoked by the test `tests/test_reporters/test_html_server::test_open_html_in_browser_watch()`
+to verify that the HTML report server can handle multiple requests when `watch=True`."""
+
+from unittest.mock import patch
+
+from python_ta.__init__ import check_all
+
+
+@patch("webbrowser.open", return_value=None)
+def open_server(mock_webbrowser_open):
+    check_all(
+        config={
+            "watch": True,
+            "pyta_port": 5008,
+        }
+    )
+
+
+open_server()
